auto-bot-producer_GPT/requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
pydantic==2.11.4
pydantic-settings==2.2.1
nest-asyncio==1.6.0
aiohttp==3.9.3
